task1_flask_firestore/requirements.txt
Flask
google-cloud-firestore
python-dotenv
